Shared create back handler
