+++ NavigationReactNative/src/TabBarItem.tsx
@@ -3,20 +3,7 @@ import { requireNativeComponent, Image, Platform, StyleSheet } from 'react-nativ
 import BackButton from './BackButton';
 import BackHandlerContext from './BackHandlerContext';
 import PrimaryStackContext from './PrimaryStackContext';
-
-var createBackHandler = () => {
-    var listeners = [];
-    var addEventListener = (eventName, handler) => {
-        if (listeners.indexOf(handler) === -1)
-            listeners.push(handler);
-        return { remove: () => removeEventListener(eventName, handler) };
-    }
-    var removeEventListener = (_, handler) => {
-        if (listeners.indexOf(handler) !== -1)
-            listeners.splice(listeners.indexOf(handler), 1);
-    }
-    return {listeners, addEventListener, removeEventListener};
-}
+import createBackHandler from './createBackHandler';
 
 class TabBarItem extends React.Component<any> {
     private backHandler: any;
@@ -26,14 +13,7 @@ class TabBarItem extends React.Component<any> {
         this.backHandler = createBackHandler();
     }
     handleBack() {
-        if (this.props.selected) {
-            var {listeners} = this.backHandler;
-            for (var i = listeners.length - 1; i >= 0; i--) {
-                if (listeners[i]())
-                    return true;
-            }
-        }
-        return false;
+        return this.props.selected && this.backHandler.handleBack();
     }
     render() {
         var {onPress, children, image, badge, index, primary, ...props} = this.props;
